Heavy refactoring of the DAP process
This refactoring brings new features and a totally brend new code base for the Data Analysis Plan.
Now the DAP is enclosed in a `data_analysis_plan` Python package that aims to be
experiment and ML/DL independent by heavy usage of Abstract Base Classes and overriding.

Code is quite well organised and heavily commented to allow an easy integration/extension of the plans
to different specific experiments.